# Conflicts:
#	metoo-rpc/service-websocket/src/main/java/com/metoo/websocket/controller/WebSocketController.java

+++ metoo-api/src/main/java/com/metoo/pojo/old/vo/MeDTO.java
@@ -14,8 +14,8 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class MeDTO implements Serializable {
     private BigDecimal balance;
-    private Integer activeIntegral;
-    private Integer psychologyIntegral;
+    private BigDecimal activeIntegral;
+    private BigDecimal psychologyIntegral;
     private String picture;
     private String name;
     private String motto;
